Random Forests Evaluation - Carbon
Looking to plot different errors for different number of estimators, sucessful just struggled with the basics of getting it to print to an image for use in report! Also working to increase the number of features used by cleaning and pickling the data set again including extra features we had previously dismissed. Further to this split the data set to a test, train and validation set so that I can complete some validation evaluation at some point soon.