Cleaning data, adding addional features and evaluating the importance using feature ranking
First stage of the work was to add additional features. These included a number of vitamins, and fat measurements that has not been used in the analysis previously.

The second stage was to rerun the algorithm with the addional features, assess the performace and evaluate which features were most important. It was found that the algorithm was over fitting when the amount of data given was too high. With around 30 different features, the model was initally overfitting and so the error on the test data was high. When the amount of data used for training was reduced, the error metrics also reduced. This was then run through the loops to look at the performance of different trees. This converged at around 10 trees as would be expected for the amount of data.

Next step is to work on bagging and boosting to see what that does for the algorithm.