Adding additional features and evaluating feature importance in the use of random forests for predicting carbon intensity.
Added in some features, mainly vitamins that we had previously dismissed. This shrunk the data set by around 100 to 1000 however led to a much improved error (absolute mean and squared included) in the model when trying to predict carbon intensity of a product. Also used sklearn to see which features were most important to making the model work effectively, i want to further analyse this but for now it seems there are 6 key features which help the model predict, 3 of which were re-added in this iteration, leading to the consolidation that it is best to include the vitamin data in the carbon predictor. Plan on next working on bagging and boosting to see how this affects model performance.  Also want to better understand the model performance by comparing to the validation, which I have set up I just havn't ran with yet. Might also be worth trying to make a classifier which predicts food type eg.bread, from carbon prediction....will discuss with team later.