Successful bagging, boosting and evaluation
Successful implementation of bagging, boosting and further evaluation. See further details in the first line of code. Gradient boosting was the best performing algorithm, this is probably to do with the fact that gradient boosting is more robust to outliers than adaboosting, due to the fact that any differentiable loss function can be used.

Gradient boosting results in slightly better performance than random forests, in terms of the errors from the model. The biggest difference between the two algorithms is that random forests use a large number of trees, using averages to predict at the end of the process. Gradient boosting also uses a large number of trees but starts the combining process at the beginning rather than doing it as the final step. Random forests build each tree independently whilst gradient boosting builds one tree at a time.