Bagging, Boosting
The work this week looked at how to improve on the random forest classification that has been carried out.

Bagging was tested as a method of comparison to the random trees. It was shown to be less efficient, and less accurate method of predicting price.

In comparison, boosted trees showed around a 5% improvement to the random forest and so this will be used in further implementation.