XGBoosting implementation
Successful and similar results to gradient boosting and random forests. Currently using default model requirements, next moving on to trying to tune the model. Also used the full data set including nulls as XG boost has the capacity to use KNN to predict for other results, this worked well for me but didnt significantly improve performance. K-fold cross validation attempted as a different method to splitting a test and train however due to the small size of the data set this didnt work well.