XG Boosting
The work looked at how to implement XG boost and minimise the error of the model. This was shown to be very successful. A number of methods were attempted.

The first stage was to implement the model with basic tuning of parameters. Options for the the objective where then looked at to see which was the best. A squared error objective was shown to be the best for the model. Given this, further tuning of parameters was looked into. This was sucessful and reduced the error further to around 0.35 in the best test.

Next stages of the work will be to look at better tuning and then to see whether CATBoost or another metod could improve the model.