XGBoosting parametre tuning
Initial set up of parameter tuning for the XGBoosting algorithm to allow for comparison to the catboost results.

Also implemented the final test and train data, need to re-run older models with this data for comparison.

Also want to compare the effect of the different sized train and test data set